+++ packages/credential-sdk/src/types/did/document/index.js
@@ -18,10 +18,12 @@ import { DidKey, DidKeys } from '../onchain/did-key';
 import { VerificationRelationship } from '../onchain/verification-relationship';
 import {
   Service,
+  DIDCommService,
   CheqdService,
-  ServiceEndpoint,
+  createServiceEndpoint,
   CheqdTestnetService,
   CheqdMainnetService,
+  ServiceEndpointId,
 } from './service-endpoint';
 import {
   VerificationMethod,
@@ -92,7 +94,20 @@ class CheqdMainnetVerificationMethods extends TypedArray {
 }
 
 export class Services extends TypedArray {
-  static Class = Service;
+  static Class = class ServiceOrDIDCommService {
+    static from(value) {
+      if (value instanceof Service || value instanceof DIDCommService) {
+        return value;
+      }
+
+      if (value.type === 'DIDCommMessaging') {
+        const se = createServiceEndpoint(value.type, value.serviceEndpoint);
+        return Service.fromServiceEndpoint(value.id, se);
+      }
+
+      return Service.from(value);
+    }
+  };
 }
 
 class CheqdServices extends TypedArray {
@@ -193,8 +208,7 @@ export class DIDDocument extends withFrom(
     }
     if (serviceEndpoints) {
       for (const [id, serviceEndpoint] of Object.entries(serviceEndpoints)) {
-        const serviceId = `${did}#${id}`;
-        doc.addServiceEndpoint(serviceId, serviceEndpoint);
+        doc.addServiceEndpoint([did, id], serviceEndpoint);
       }
     }
 
@@ -447,8 +461,11 @@ export class CheqdDIDDocument extends TypedStruct {
     // Convert CheqdService objects to Service objects
     const convertedServices = [...service].map((cheqdService) => {
       // Create a ServiceEndpoint object from the CheqdService data
-      const serviceEndpoint = new ServiceEndpoint(cheqdService.serviceType, cheqdService.serviceEndpoint);
-      return Service.fromServiceEndpoint(cheqdService.id, serviceEndpoint);
+      const serviceEndpointArray = cheqdService.serviceEndpoint.toJSON();
+      const serviceEndpoint = createServiceEndpoint(cheqdService.serviceType, serviceEndpointArray);
+      // Convert the Cheqd-specific ID to a generic ID
+      const genericId = ServiceEndpointId.from(cheqdService.id.toJSON());
+      return Service.fromServiceEndpoint(genericId, serviceEndpoint);
     });
 
     const assertionMethodOffchainKeys = [...assertionMethod].filter(